refactor: 기존에 존재하던 School 조회를 Admin으로 변경
