+++ backend/src/main/java/com/festago/admin/presentation/v1/AdminSchoolV1Controller.java
@@ -2,18 +2,26 @@ package com.festago.admin.presentation.v1;
 
 import com.festago.admin.presentation.v1.dto.SchoolV1CreateRequest;
 import com.festago.admin.presentation.v1.dto.SchoolV1UpdateRequest;
+import com.festago.common.aop.ValidPageable;
+import com.festago.common.querydsl.SearchCondition;
+import com.festago.school.application.AdminSchoolV1QueryService;
 import com.festago.school.application.SchoolCommandService;
 import com.festago.school.application.SchoolDeleteService;
+import com.festago.school.presentation.v1.dto.SchoolV1Response;
 import io.swagger.v3.oas.annotations.Hidden;
 import java.net.URI;
 import lombok.RequiredArgsConstructor;
+import org.springframework.data.domain.Page;
+import org.springframework.data.domain.Pageable;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.DeleteMapping;
+import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PatchMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
+import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 @RestController
@@ -24,6 +32,7 @@ public class AdminSchoolV1Controller {
 
     private final SchoolCommandService schoolCommandService;
     private final SchoolDeleteService schoolDeleteService;
+    private final AdminSchoolV1QueryService schoolQueryService;
 
     @PostMapping
     public ResponseEntity<Void> createSchool(
@@ -51,5 +60,24 @@ public class AdminSchoolV1Controller {
         schoolDeleteService.deleteSchool(schoolId);
         return ResponseEntity.noContent()
             .build();
+    }
+
+    @GetMapping
+    @ValidPageable(maxSize = 20)
+    public ResponseEntity<Page<com.festago.school.presentation.v1.dto.SchoolV1Response>> findAllSchools(
+            @RequestParam(defaultValue = "") String searchFilter,
+            @RequestParam(defaultValue = "") String searchKeyword,
+            Pageable pageable
+    ) {
+        return ResponseEntity.ok()
+                .body(schoolQueryService.findAll(new SearchCondition(searchFilter, searchKeyword, pageable)));
+    }
+
+    @GetMapping("/{schoolId}")
+    public ResponseEntity<SchoolV1Response> findSchoolById(
+            @PathVariable Long schoolId
+    ) {
+        return ResponseEntity.ok()
+                .body(schoolQueryService.findById(schoolId));
     }
 }
