+++ backend/src/main/java/com/festago/school/application/AdminSchoolV1QueryService.java
@@ -4,7 +4,7 @@ import com.festago.common.exception.ErrorCode;
 import com.festago.common.exception.NotFoundException;
 import com.festago.common.querydsl.SearchCondition;
 import com.festago.school.presentation.v1.dto.SchoolV1Response;
-import com.festago.school.repository.SchoolV1QueryDslRepository;
+import com.festago.school.repository.AdminSchoolV1QueryDslRepository;
 import lombok.RequiredArgsConstructor;
 import org.springframework.data.domain.Page;
 import org.springframework.stereotype.Service;
@@ -13,9 +13,9 @@ import org.springframework.transaction.annotation.Transactional;
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
-public class SchoolV1QueryService {
+public class AdminSchoolV1QueryService {
 
-    private final SchoolV1QueryDslRepository schoolQueryDslRepository;
+    private final AdminSchoolV1QueryDslRepository schoolQueryDslRepository;
 
     public Page<SchoolV1Response> findAll(SearchCondition searchCondition) {
         return schoolQueryDslRepository.findAll(searchCondition);
