+++ backend/src/test/java/com/festago/acceptance/steps/SchoolStepDefinitions.java
@@ -53,7 +53,8 @@ public class SchoolStepDefinitions {
         return RestAssured.given()
             .contentType(ContentType.JSON)
             .queryParams(Map.of("searchFilter", "name", "searchKeyword", src))
-            .get("/api/v1/schools")
+            .cookie("token", cucumberClient.getToken())
+            .get("/admin/api/v1/schools")
             .then()
             .log().ifError()
             .statusCode(200)
